fluke-dmm: Move device-specific DM handling
With the exception of the Fluke 28x series, device-specific DM
handling is currently handled in protocol.c. Move this to separate
files for clarity.
